Rename ops package to operations
