+++ src/main/java/backend/control/operations/OpenRepoOp.java
@@ -1,4 +1,4 @@
-package backend.control.ops;
+package backend.control.operations;
 
 import static util.Futures.chain;
 
@@ -7,26 +7,26 @@ import java.util.concurrent.CompletableFuture;
 import backend.RepoIO;
 import backend.resource.Model;
 
-public class UpdateModelOp implements RepoOp<Model> {
+public class OpenRepoOp implements RepoOp<Model> {
 
-    private final Model oldModel;
+    private final String repoId;
     private final RepoIO repoIO;
     private final CompletableFuture<Model> result;
 
-    public UpdateModelOp(Model oldModel, RepoIO repoIO, CompletableFuture<Model> result) {
-        this.oldModel = oldModel;
+    public OpenRepoOp(String repoId, RepoIO repoIO, CompletableFuture<Model> result) {
+        this.repoId = repoId;
         this.repoIO = repoIO;
         this.result = result;
     }
 
     @Override
     public String repoId() {
-        return oldModel.getRepoId();
+        return repoId;
     }
 
     @Override
     public CompletableFuture<Model> perform() {
-        return repoIO.updateModel(oldModel)
+        return repoIO.openRepository(repoId)
             .thenApply(chain(result));
     }
 }
