+++ src/main/java/backend/control/operations/ReplaceIssueLabelsOp.java
@@ -1,4 +1,4 @@
-package backend.control.ops;
+package backend.control.operations;
 
 import backend.RepoIO;
 import backend.resource.TurboIssue;
