+++ src/main/java/backend/control/operations/UpdateModelOp.java
@@ -1,4 +1,4 @@
-package backend.control.ops;
+package backend.control.operations;
 
 import static util.Futures.chain;
 
@@ -7,26 +7,26 @@ import java.util.concurrent.CompletableFuture;
 import backend.RepoIO;
 import backend.resource.Model;
 
-public class OpenRepoOp implements RepoOp<Model> {
+public class UpdateModelOp implements RepoOp<Model> {
 
-    private final String repoId;
+    private final Model oldModel;
     private final RepoIO repoIO;
     private final CompletableFuture<Model> result;
 
-    public OpenRepoOp(String repoId, RepoIO repoIO, CompletableFuture<Model> result) {
-        this.repoId = repoId;
+    public UpdateModelOp(Model oldModel, RepoIO repoIO, CompletableFuture<Model> result) {
+        this.oldModel = oldModel;
         this.repoIO = repoIO;
         this.result = result;
     }
 
     @Override
     public String repoId() {
-        return repoId;
+        return oldModel.getRepoId();
     }
 
     @Override
     public CompletableFuture<Model> perform() {
-        return repoIO.openRepository(repoId)
+        return repoIO.updateModel(oldModel)
             .thenApply(chain(result));
     }
 }
